requirements.txt
python-dotenv
nltk
unstructured
unstructured[pdf]
langchain
langchain-community
langchain-huggingface
langchain-chroma
langchain-text-splitters
langchain-groq
transformers
sentence-transformers
httpx